requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
rank-bm25>=0.2.2
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0
nltk>=3.8.1
